build_zol.sh: Only git push if Debian GNU/Linux (others not yet supported).

+++ build_zol.sh
@@ -291,8 +291,10 @@ fi
 
 # Push our changes to GitHub
 if [ -e "/etc/debian_version" ]; then
-    git push pkg-${APP} --force --all
-    git push pkg-${APP} --force --tags
+    if echo "${DIST}" | grep -Eq "wheezy|jessie|sid"; then
+	git push pkg-${APP} --force --all
+	git push pkg-${APP} --force --tags
+    fi
 fi
 
 # Record changes
